backend/app/requirements.txt
fastapi[all] 
uvicorn 
pydantic 
pymongo 
motor 
python-dotenv 
jose 
fastapi-cors
passlib[bcrypt] 
python-jose[cryptography] 
requests
httpx
google-auth
google-auth-oauthlib 
google-auth-httplib2
pydantic-settings
pandas 
openpyxl